LayoutTests/http/tests/security/dataURL/xss-DENIED-from-data-url-sub-frame-to-data-url-sub-frame-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,%3Chtml%3E%3Cbody%3E%3Cp%3EInner%20iframe.%3C/p%3E%3Cp%20id='accessMe'%3EPass:%20Cross%20frame%20access%20from%20a%20sibling%20data:%20URL%20was%20denied.%3C/p%3E%3Ciframe%3E%3C/iframe%3E%3C/body%3E%3C/html%3E from frame with URL data:text/html,%3Chtml%3E%3Cbody%3E%3Cp%3EInner%20iframe.%3C/p%3E%3Cscript%3Eparent.frames['firstSubFrame'].document.getElementById('accessMe').innerHTML%20=%20'FAIL:%20Cross%20frame%20access%20from%20a%20sibling%20data:%20URL%20was%20allowed.';%3C/script%3E%3Ciframe%3E%3C/iframe%3E%3C/body%3E%3C/html%3E. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: TypeError: Undefined value
This tests that a data: URL subframe can't access a sibling data: URL subframe.

  


--------
Frame: 'firstSubFrame'
--------
Inner iframe.

Pass: Cross frame access from a sibling data: URL was denied.



--------
Frame: '<!--framePath //firstSubFrame/<!--frame0-->-->'
--------


--------
Frame: 'secondSubFrame'
--------
Inner iframe.



--------
Frame: '<!--framePath //secondSubFrame/<!--frame0-->-->'
--------
